[docs-infra] Move ads to the @mui/docs package
