+++ packages/mui-docs/src/Ad/Ad.tsx
@@ -3,10 +3,10 @@ import PropTypes from 'prop-types';
 import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
 import Paper from '@mui/material/Paper';
-import AdCarbon from 'docs/src/modules/components/AdCarbon';
-import AdInHouse from 'docs/src/modules/components/AdInHouse';
-import { AdContext, adShape } from 'docs/src/modules/components/AdManager';
-import { useTranslate } from '@mui/docs/i18n';
+import { useTranslate } from '../i18n';
+import AdCarbon from './AdCarbon';
+import AdInHouse from './AdInHouse';
+import { AdContext, adShape } from './AdManager';
 
 function PleaseDisableAdblock() {
   const t = useTranslate();
