+++ packages/mui-docs/src/Ad/AdInHouse.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import AdDisplay, { Ad } from 'docs/src/modules/components/AdDisplay';
+import AdDisplay, { AdParameters } from './AdDisplay';
 
-export default function AdInHouse(props: { ad: Omit<Ad, 'poweredby' | 'label'> }) {
+export default function AdInHouse(props: { ad: Omit<AdParameters, 'poweredby' | 'label'> }) {
   const { ad } = props;
 
   return <AdDisplay ad={{ poweredby: 'MUI', label: `in-house-${ad.name}`, ...ad }} />;
