+++ packages/mui-docs/src/Ad/index.ts
@@ -1,0 +1,3 @@
+export * from './Ad';
+export * from './AdManager';
+export { default as AdGuest } from './AdGuest';
